Move glamorous containers to separate files
