+++ src/App.js
@@ -1,11 +1,13 @@
 import React from 'react';
-import glamorous from 'glamorous';
 
 import api from './api';
 
 import Layout from './components/Layout/Layout';
 import Header from './components/Header/Header';
+import Content from './components/Content/Content';
 import Footer from './components/Footer/Footer';
+import Link from './components/styled/Link';
+import Paragraph from './components/styled/Paragraph';
 import SearchForm from './components/SearchForm/SearchForm';
 import Collections from './components/Collections/Collections';
 
@@ -14,18 +16,19 @@ const App = () => (
     <Header>
       <SearchForm />
     </Header>
-    <Container>
+    <Content>
       <Collections images={api} />
-    </Container>
-    <Footer />
+    </Content>
+    <Footer>
+      <Paragraph>
+        The source code is licensed under MIT. See{' '}
+        <Link href="https://github.com/synthetiquely/collections">
+          public repository
+        </Link>{' '}
+        for more details.
+      </Paragraph>
+    </Footer>
   </Layout>
 );
-
-const Container = glamorous.main({
-  gridArea: 'content',
-  gridColumn: '1 / -1',
-  gridRow: '2 / 3',
-  backgroundColor: '#f6f5f3',
-});
 
 export default App;
