+++ src/components/Content/Content.js
@@ -1,0 +1,10 @@
+import glamorous from 'glamorous';
+
+const Content = glamorous.main({
+  gridArea: 'content',
+  gridColumn: '1 / -1',
+  gridRow: '2 / 3',
+  backgroundColor: '#f6f5f3',
+});
+
+export default Content;
