+++ src/components/Footer/Footer.js
@@ -1,7 +1,4 @@
-import React from 'react';
 import glamorous from 'glamorous';
-import Link from '../styled/Link';
-import Paragraph from '../styled/Paragraph';
 
 const Footer = glamorous.footer({
   gridArea: 'footer',
@@ -13,14 +10,4 @@ const Footer = glamorous.footer({
   borderTop: '1px solid rgba(0,0,0,.1)',
 });
 
-export default () => (
-  <Footer>
-    <Paragraph>
-      The source code is licensed under MIT. See{' '}
-      <Link href="https://github.com/synthetiquely/collections">
-        public repository
-      </Link>{' '}
-      for more details.
-    </Paragraph>
-  </Footer>
-);
+export default Footer;
