+++ src/components/Header/Header.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import glamorous from 'glamorous';
 
 const Header = glamorous.header({
@@ -11,4 +10,4 @@ const Header = glamorous.header({
   borderBottom: '1px solid rgba(0,0,0,.1)',
 });
 
-export default props => <Header>{props.children}</Header>;
+export default Header;
