+++ src/components/View/View.js
@@ -1,67 +1,25 @@
 import React, { Component } from 'react';
-import glamorous from 'glamorous';
-import { SIZE_RATIO } from '../../constants';
+import ImageContainer from './ImageContainer';
+import Image from './Image';
 
-const ImageContainer = glamorous.div(
-  {
-    margin: '2.5px',
-    cursor: 'pointer',
-    willChange: 'width, flexGrow',
-    ':hover': {
-      boxShadow: '1px 1px 1px #ccc',
-    },
-  },
-  ({ height, width }) => {
-    if (height && width) {
-      const size = Math.floor(width * SIZE_RATIO / height);
-
-      return {
-        flexGrow: `${size}`,
-        width: `${size}px`,
-      };
-    }
-    return {};
-  },
-);
-
-const Image = glamorous.img(
-  {
-    opacity: '0',
-    width: '100%',
-    verticalAlign: 'bottom',
-    willChange: 'opacity',
-    transition: 'opacity 2s',
-  },
-  ({ loaded }) => {
-    if (loaded) {
-      return {
-        opacity: '1',
-      };
-    }
-    return {};
-  },
-);
-
-export default class View extends Component {
+class View extends Component {
   constructor(props) {
     super(props);
     this.state = { loaded: false };
     this.width = 0;
     this.height = 0;
-    this.img = null;
     this.onLoad = this.onLoad.bind(this);
     this.onError = this.onError.bind(this);
-    this.setRef = this.setRef.bind(this);
   }
 
-  onLoad() {
+  onLoad(event) {
     /**
      * Due to the fact that img width and height will be known in advance
      * as soon as we bring API to our app,
      * we could remove this lines of code.
      */
-    this.width = this.img.naturalWidth || this.img.width;
-    this.height = this.img.naturalHeight || this.img.height;
+    this.width = event.target.naturalWidth || event.target.width;
+    this.height = event.target.naturalHeight || event.target.height;
     this.setState({
       loaded: true,
     });
@@ -69,10 +27,6 @@ export default class View extends Component {
 
   // @TODO: handle error case
   onError() {}
-
-  setRef(img) {
-    this.img = img;
-  }
 
   render() {
     const { src, title, onClick } = this.props;
@@ -82,7 +36,6 @@ export default class View extends Component {
         <Image
           src={src}
           alt={title}
-          innerRef={this.setRef}
           loaded={this.state.loaded}
           onLoad={this.onLoad}
           onError={this.onError}
@@ -91,3 +44,5 @@ export default class View extends Component {
     );
   }
 }
+
+export default View;
