+++ src/components/styled/Form.js
@@ -1,0 +1,8 @@
+import glamorous from 'glamorous';
+
+const Form = glamorous.form({
+  display: 'flex',
+  flexDirection: 'row',
+});
+
+export default Form;
